requirements.txt
bcrypt==4.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==43.0.3
idna==3.10
mysql-connector-python==9.1.0
numpy==2.0.2
pandas==2.2.3
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
swat==1.15.0
tzdata==2024.2
urllib3==2.3.0
xmltodict==0.14.2
